Update dtype convertion in the `_temperature_scaling` function. Update input dtype to `numpy.float64` in `test_tempereature_scaling_input_validation` function.

+++ sklearn/tests/test_calibration.py
@@ -497,8 +497,8 @@ def test_temperature_scaling(clf, n_classes):
 
 
 def test_temperature_scaling_input_validation():
-    # Check that _TemperatureScaling can handle 2darray with only 1 feature
-    X = np.arange(10)
+    # Check that _TemperatureScaling can handle 2d-array with only 1 feature
+    X = np.arange(10).astype(np.float64)
     X_2d = X.reshape(-1, 1)
     y = np.random.randint(0, 2, size=X.shape[0])
